pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy==1.24.3",
    "pandas==2.0.3",
    "python-deriv-api>=0.1.6",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "tensorflow>=2.14.0",
    "websockets>=10.3",
]
